% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newton_abs_debug_solver}
\alias{newton_abs_debug_solver}
\title{Newton solver with more information.}
\usage{
newton_abs_debug_solver(L, v, l1, l2, tol, max_iters, smart_init = TRUE)
}
\arguments{
\item{L}{vector representing a diagonal PSD matrix.
Must have max(L + s) > 0.
L.size() <= buffer1.size().}

\item{v}{any vector.}

\item{l1}{L2-norm penalty. Must be >= 0.}

\item{l2}{L2 penalty. Must be >= 0.}

\item{tol}{Newton's method tolerance of closeness to 0.}

\item{max_iters}{maximum number of iterations of Newton's method.}
}
\description{
Newton solver with more information.
}
